:up: Improve auto-fragments
By default, when a slide contains indexed fragments and non-indexed ones, the non-indexed appear __after__ the indexed ones. This makes sense because indexes start from 0 and indexing under that is not always possible, depending on the user-set indexes.

This commit adds a class `.auto-fragments-first` that adds fragments indexes on all non-indexed elements by order of appearance. This makes sure Reveal doesn't try to index them itself and allows the user to specify bigger indexes for the manually indexed fragments so that they appear after the auto-indexed fragments.